Plan 9 from Bell Labs 2004-08-06

+++ sys/src/cmd/5i/5i.c
@@ -6,7 +6,7 @@
 
 char*	file = "5.out";
 int	datasize;
-int	textbase;
+ulong	textbase;
 Biobuf	bp, bi;
 Fhdr	fhdr;
